IN-664 Refactoring geotag_service.rb and nlp_api_service.rb
Merge nlp_api_service.rb into nlp/api.rb, where it belongs.
