--- back/engines/commercial/geographic_dashboard/app/services/geographic_dashboard/nlp_api_service.rb
@@ -1,16 +1,0 @@
-module GeographicDashboard
-  class NLPApiService < ::NLP::Api
-    def geotag(tenant_id, text, locale, options = {})
-      body = options.merge(text: text, locale: locale)
-      resp = post(
-        "/v1/tenants/#{tenant_id}/geotagging",
-        body: body.to_json,
-        headers: authorization_header.merge('Content-Type' => 'application/json'),
-        timeout: LONG_TIMEOUT
-      )
-      raise ClErrors::TransactionError.new(error_key: resp['code']) unless resp.success?
-
-      resp.parsed_response['data']
-    end
-  end
-end
