+++ back/engines/commercial/nlp/app/services/nlp/api.rb
@@ -164,6 +164,25 @@ module NLP
     end
     # rubocop:enable Metrics/MethodLength
 
+
+    # @param [String] tenant_id
+    # @param [String] text
+    # @param [String] locale
+    def geotag(tenant_id, text, locale, options = {})
+      body = options.merge(text: text, locale: locale)
+
+      resp = post(
+        "/v1/tenants/#{tenant_id}/geotagging",
+        body: body.to_json,
+        headers: authorization_header.merge('Content-Type' => 'application/json'),
+        timeout: LONG_TIMEOUT
+      )
+      raise ClErrors::TransactionError.new(error_key: resp['code']) unless resp.success?
+
+      resp.parsed_response['data']
+    end
+
+
     def authorization_header
       { 'Authorization' => "Token #{@authorization_token}"}
     end
